(fixup) move version string builder methods to a util
